Note museums that are set up as users rather than orgs in GitHub
I realized a number of museums are set up in GitHub in user accounts rather than as orgs. As a result I was missing them when pulling data from GitHub's API. However, the key org account data is available through the users API as well so by pulling from there I could get everyone whether they were an org or an individual user. Thought I'd make a note of it in this list as well though.