Add script to capture data on museums from GitHub API
This is to be wired into layouts/data/barchart.html where currently the years orgs joined is hard-coded in. Also should be used to populate the owners dropdown list in static/js/barchart.js, which too is hard-coded in. For this last, the museums in the list can only be those with public repositories for the visualization to function properly.